+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.8.1 (2024-10-11)
+
+### Fixed
+
+- cost model for `integerToBytestring` and `bytestringToInteger`
+- flat encoding of SIR `Data` was wrong for `Map` and `List` types
+
 ## 0.8.0 (2024-10-03)
 
 ### Added